move run_all_simulations to htsohm.py
